Renamed artist component folders
