+++ public/index.html
@@ -19,7 +19,7 @@
       href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700&display=swap"
       rel="stylesheet"
     />
-    <title>Desafio Front-end</title>
+    <title>Plurall - Music</title>
   </head>
 
   <body>
