+++ src/components/Artist/ArtistDetailedCard.module.scss
