+++ src/components/Artist/index.js
