+++ src/components/SearchArtist/ArtistSimplifiedCard.jsx
