+++ src/components/SearchArtist/ArtistSimplifiedCard.module.scss
