+++ src/components/SearchArtist/ArtistsList.module.scss
