+++ src/components/SearchArtist/LoadingArtistSmallCard.jsx
