+++ src/components/SearchArtist/LoadingArtistSmallCard.module.scss
