+++ src/components/SearchArtist/LoadingArtistsList.jsx
