+++ src/components/SearchArtist/LoadingArtistsList.module.scss
