+++ src/components/SearchArtist/SearchRemainingLetters.module.scss
