+++ src/components/SearchArtist/index.js
